% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarifyCode.R
\name{clarifyCode}
\alias{clarifyCode}
\title{LLM: Clarify Code}
\usage{
clarifyCode(
  code = clipr::read_clip(allow_non_interactive = TRUE),
  PERPLEXITY_API_KEY = Sys.getenv("PERPLEXITY_API_KEY"),
  modelSelection = c("codellama-34b-instruct", "llama-2-70b-chat", "mistral-7b-instruct",
    "mixtral-8x7b-instruct", "pplx-7b-chat", "pplx-70b-chat", "pplx-7b-online",
    "pplx-70b-online"),
  systemRole = "You are a helpful assistant with extensive programming skills.",
  maxTokens = 265,
  temperature = 1,
  top_p = NULL,
  top_k = 100,
  presence_penalty = 0,
  frequency_penalty = NULL,
  proxy = NULL,
  returnType = 1
)
}
\arguments{
\item{code}{The code to be explained by LLM. If not provided, it will use what's copied on
the clipboard.}

\item{PERPLEXITY_API_KEY}{PERPLEXITY API key.}

\item{modelSelection}{model choice. Default is mistral-7b-instruct.}

\item{systemRole}{Role for model. Default is: "You are a helpful assistant
with extensive knowledge of R programming."}

\item{maxTokens}{The maximum integer of completion tokens returned by API.}

\item{temperature}{The amount of randomness in the response,
valued between 0 inclusive and 2 exclusive. Higher values are more random,
and lower values are more deterministic. Set either temperature or top_p.}

\item{top_p}{Nucleus sampling threshold, valued between 0 and 1 inclusive.}

\item{top_k}{The number of tokens to keep for highest top-k filtering,
specified as an integer between 0 and 2048 inclusive.
If set to 0, top-k filtering is disabled.}

\item{presence_penalty}{A value between -2.0 and 2.0.
Positive values penalize new tokens based on whether they appear in the text
so far, increasing the model's likelihood to talk about new topics.
Incompatible with frequency_penalty.}

\item{frequency_penalty}{A multiplicative penalty greater than 0.
Values greater than 1.0 penalize new tokens based on their existing
frequency in the text so far, decreasing the model's likelihood to repeat
the same line verbatim. A value of 1.0 means no penalty.}

\item{proxy}{Default value is NULL.}

\item{returnType}{Default is 1, which cats the output, type 2 is unchanged
and type 3 returns the output to the clipboard and returns TRUE.}
}
\value{
A character value with the response generated by LLM.
}
\description{
LLM: Clarify Code
}
\examples{
\dontrun{
cat(clarifyCode("z <- function(x) scale(x)^2"))
}

}
