% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{execAddin}
\alias{execAddin}
\title{Run a Large Language Model as RStudio add-in}
\usage{
execAddin(FUN)
}
\arguments{
\item{FUN}{The function to be executed.}
}
\description{
Run a Large Language Model as RStudio add-in
}
