% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerplexR-package.R
\docType{package}
\name{PerplexR-package}
\alias{PerplexR}
\alias{PerplexR-package}
\title{PerplexR: A coding assistant using Perplexity's LLMs.}
\description{
'Perplexity's 'LLM' \url{https://perplexity.ai/} coding assistant for 'RStudio' or any other IDE. A set of functions and 'RStudio' addins that aim to help developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GabrielKaiserQFin/PerplexR}
  \item Report bugs at \url{https://github.com/GabrielKaiserQFin/PerplexR/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Kaiser \email{quantresearch.gk@gmail.com}

}
\keyword{LLAMA,}
\keyword{LLM,}
\keyword{Mistral,}
\keyword{Openhermes,}
\keyword{PPLX,}
\keyword{annotate;}
\keyword{debug,}
\keyword{document,}
\keyword{optimize,}
\keyword{translate,}
