% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseParser.R
\name{responseParser}
\alias{responseParser}
\title{Parse Perplexity API Response}
\usage{
responseParser(raw)
}
\arguments{
\item{raw}{The raw object returned by the Perplexity API.}
}
\value{
Returns a character vector containing the text content of the response.
}
\description{
Takes the raw response from the Perplexity API and extracts the text content from it.
}
