% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseReturn.R
\name{responseReturn}
\alias{responseReturn}
\title{responseReturn}
\usage{
responseReturn(raw)
}
\arguments{
\item{raw}{the chatresponse to return}
}
\value{
A character value with the response generated by Large Language Model.
}
\description{
responseReturn
}
