% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerplexR-package.R
\docType{package}
\name{PerplexR-package}
\alias{PerplexR}
\alias{PerplexR-package}
\title{PerplexR: A Coding Assistant using Perplexity's LLMs}
\description{
A coding assistant using Perplexity's LLMs \url{https://perplexity.ai/} API. A set of functions and 'RStudio' Addins that aim to help R developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GabrielKaiserQFin/PerplexR}
  \item Report bugs at \url{https://github.com/GabrielKaiserQFin/PerplexR/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Kaiser \email{quantresearch.gk@gmail.com}

}
\keyword{LLAMA}
\keyword{LLM}
\keyword{Mistral}
\keyword{Openhermes}
\keyword{PPLX}
\keyword{annotate;}
\keyword{debug}
\keyword{document}
\keyword{optimize}
\keyword{translate}
