% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexR-package.R
\docType{package}
\name{perplexR-package}
\alias{perplexR}
\alias{perplexR-package}
\title{perplexR: A Coding Assistant using Perplexity's Large Language Models}
\description{
A coding assistant using Perplexity's Large Language Models \url{https://www.perplexity.ai/} API. A set of functions and 'RStudio' add-ins that aim to help R developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GabrielKaiserQFin/perplexR}
  \item Report bugs at \url{https://github.com/GabrielKaiserQFin/perplexR/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Kaiser \email{quantresearch.gk@gmail.com}

}
\keyword{LLAMA}
\keyword{Language}
\keyword{Large}
\keyword{Mistral}
\keyword{Model}
\keyword{Openhermes}
\keyword{PPLX}
\keyword{annotate;}
\keyword{debug}
\keyword{document}
\keyword{optimize}
\keyword{translate}
