% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseReturn.R
\name{responseReturn}
\alias{responseReturn}
\title{responseReturn}
\usage{
responseReturn(raw, type)
}
\arguments{
\item{raw}{the chatresponse to return}

\item{type}{Default is 1, which cats the output, type 2 is unchanged and type 3 returns the output to the clipboard and returns TRUE}
}
\value{
A character value with the response generated by LLM.
}
\description{
responseReturn
}
